requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
cachetools==5.5.0
catalogue==2.0.10
catboost==1.2.8
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
faiss-cpu==1.11.0.post1
Flask==3.1.1
flatbuffers==24.3.25
fonttools==4.55.5
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.68.1
grpcio-status==1.62.3
gspread==6.2.1
gTTS==2.5.4
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.4.8
language_data==1.3.0
libclang==18.1.1
lightgbm==4.6.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
MouseInfo==0.1.3
murmurhash==1.0.13
namex==0.0.8
narwhals==1.15.2
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
openai==1.97.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.13.1
orjson==3.11.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
playsound==1.3.0
plotly==6.2.0
preshed==3.0.10
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==4.25.8
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.1
pypdf==4.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.1
seaborn==0.13.2
selenium==4.34.2
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
slack_bolt==1.21.3
slack_sdk==3.36.0
smart_open==7.3.0.post1
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
streamlit==1.37.1
streamlit-option-menu==0.4.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xgboost==3.0.2
yt-dlp==2025.5.22
zstandard==0.23.0
